Add abstract SoftConstraint
