+++ app/src/main/java/io/ipoli/android/app/scheduling/constraints/MorningConstraint.java
@@ -1,10 +1,6 @@
 package io.ipoli.android.app.scheduling.constraints;
 
-import io.ipoli.android.app.scheduling.DailySchedule;
 import io.ipoli.android.app.scheduling.Task;
-import io.ipoli.android.app.scheduling.distributions.DiscreteDistribution;
-import io.ipoli.android.app.scheduling.distributions.FlatPeakDiscreteDistribution;
-import io.ipoli.android.app.scheduling.distributions.UniformDiscreteDistribution;
 import io.ipoli.android.app.utils.TimePreference;
 
 import static io.ipoli.android.app.utils.Time.h2Min;
@@ -13,30 +9,18 @@ import static io.ipoli.android.app.utils.Time.h2Min;
  * Created by Venelin Valkov <venelin@curiousily.com>
  * on 4/23/17.
  */
-public class MorningConstraint implements Constraint {
+public class MorningConstraint extends SoftConstraint {
 
     public static final int MORNING_START = h2Min(6);
     public static final int MORNING_END = h2Min(12);
+
+    public MorningConstraint() {
+        super(MORNING_START, MORNING_END);
+    }
 
     @Override
     public boolean shouldApply(Task task) {
         return task.getStartTimePreference() == TimePreference.MORNING;
     }
-
-    @Override
-    public DiscreteDistribution apply(DailySchedule schedule) {
-        int startMinute = schedule.getStartMinute();
-        int endMinute = schedule.getEndMinute();
-
-        if(startMinute > MORNING_END || endMinute < MORNING_START) {
-            return UniformDiscreteDistribution.create(schedule.getSlotCount());
-        }
-
-        int slopeWidth = schedule.getSlotCountBetween(0, 30);
-        int peakWidth = schedule.getSlotCountBetween(Math.max(MORNING_START, startMinute), Math.min(MORNING_END, endMinute));
-        int peakStart = schedule.getSlotForMinute(Math.max(MORNING_START, startMinute));
-        return FlatPeakDiscreteDistribution.create(peakStart, peakWidth, 100, slopeWidth, schedule.getSlotCount());
-    }
-
 
 }
